requirements.txt
pandas
sklearn
tqdm
sqlalchemy
networkx
neo4j
setuptools
pyyaml
jinja2
psycopg2
Sphinx